Bugfix in lib/opts
Previously if there were no flags/positional arguments the opts function
would set the global variables to the empty string, when they should
have been unset. This caused issues when, for example, no flags were
supplied and programs saw the empty string as a stray positional
argument.

Additionally the name "rest" was changed to "pos" to make the meaning
clearer.