Port to matrix chat
Remove irc-specific and otherwise broken things in bin and lib,
and port init to matrix_client. Also add some commands and fix some old
ones.

Add a Dockerfile to set up the environment in a consistent way. This is
helpful in particular because Python3 IO encodes/decodes based on the
system locale.